backend/requirements.txt
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
flask-cors==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
requests==2.32.4
urllib3==2.5.0
Werkzeug==3.1.3
xmltodict
pdfplumber
PyPDF2
pymongo
python-dotenv
PyJWT
# openai-whisper
SpeechRecognition
pydub
sqlalchemy
psycopg2-binary
python-dateutil
pytz
alembic
apscheduler
python-dotenv
pandas
PyYAML
XlsxWriter
faster-whisper
python-magic
redis
cryptography
passlib[bcrypt]
lxml